refactor related space into separate widget
